[EDT-1538] PR comments
