+++ packages/sdk/src/controllers/ConfigurationController.ts
@@ -8,7 +8,7 @@ import { StudioOptionsDeltaUpdate, WellKnownConfigurationKeys } from '../types/C
  * Javascript actions running in the editor.
  * Methods inside this controller can be called by `window.SDK.configuration.{method-name}`
  */
-export class ConfigurationController {
+export abstract class ConfigurationController {
     /**
      * @ignore
      */
@@ -17,7 +17,7 @@ export class ConfigurationController {
     /**
      * @ignore
      */
-    constructor(editorAPI: EditorAPI) {
+    protected constructor(editorAPI: EditorAPI) {
         this.#editorAPI = editorAPI;
     }
 
@@ -26,36 +26,36 @@ export class ConfigurationController {
      * this method returns an error. The key cannot be null.
      * @returns value for a given configuration key
      */
-    getValue = async (key: WellKnownConfigurationKeys | string) => {
+    async getValue(key: WellKnownConfigurationKeys | string) {
         const res = await this.#editorAPI;
         return res.getConfigValue(key).then((result) => getEditorResponseData<string>(result));
-    };
+    }
 
     /**
      * This method sets, or overrides the value for a given key. Null values are not allowed for both key and value,
      * using them will result in an error.
      * @returns
      */
-    setValue = async (key: WellKnownConfigurationKeys | string, value: string) => {
+    async setValue(key: WellKnownConfigurationKeys | string, value: string) {
         const res = await this.#editorAPI;
         return res.setConfigValue(key, value).then((result) => getEditorResponseData<null>(result));
-    };
+    }
 
     /**
      * This method updates the studio options. Only defined options are updated, null values are ignored.
      * @returns
      */
-    updateStudioOptions = async (options: StudioOptionsDeltaUpdate) => {
+    async updateStudioOptions(options: StudioOptionsDeltaUpdate) {
         const res = await this.#editorAPI;
         return res.updateStudioOptions(JSON.stringify(options)).then((result) => getEditorResponseData<null>(result));
-    };
+    }
 
     /**
      * This method returns the engine session id. This id is unique for this engine session
      * @returns the engine session id
      */
-    getEngineSessionId = async () => {
+    async getEngineSessionId() {
         const res = await this.#editorAPI;
         return res.getEngineSessionId().then((result) => getEditorResponseData<string>(result));
-    };
+    }
 }
